[fix] Modalコンポーネントの構成変更 & 軽くレスポンシブ化
修正commitの細分化を忘れる…
