+++ view-user/src/components/common/Button/Button.module.css
@@ -4,9 +4,10 @@
   color: #333333;
   text-align: center;
   cursor: pointer;
-  position: absolute;
   align-items: center;
   display: flex;
+  flex-flow: column;
+  justify-content: center;
   box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
 }
 
@@ -55,3 +56,7 @@
   color: var(--primary);
   transition: 0.25s;
 }
+
+@media screen and (max-width: 900px){
+  .primary {display:none}
+  }
