+++ view-user/src/components/common/Header/Header.module.css
@@ -1,6 +1,6 @@
 /* Header */
 .container {
-    height: 6rem;
+    height: 6%;
     width: 100vw;
     background-color: #D9D9D9;
     border-bottom: 0.3rem solid #333;
@@ -23,22 +23,3 @@
     display: inline-block;
 }
 
-/* Modal */
-.modalContent {
-    display: flex;
-    flex-flow: column;
-    justify-content: center;
-    align-items: center;
-    color: #D9D9D9;
-    font-size: 2rem;
-    font-weight: 700;
-    font-feature-settings: "palt" 1;
-}
-
-.modalContent p {
-    color: #D9D9D9;
-    font-size: 14rem;
-    font-weight: 700;
-    margin-top: -3rem;
-    margin-bottom: -3rem;
-}
