+++ view-user/src/components/common/Header/Header.tsx
@@ -1,30 +1,16 @@
-import React from 'react'
-import { ReactNode } from "react"
-import styles from "./Header.module.css"
-import { Modal,Button } from "@/components/common";
+import React, { ReactNode } from "react";
+import styles from "./Header.module.css";
 
 interface HeaderProps {
-  user: string;
+    children: ReactNode;
 }
 
-const Header = (props: HeaderProps) =>  {
+const Header = ({ children }: HeaderProps) => {
     return (
         <div className={styles.container}>
-            <div className={styles.main}>
-                <div className={styles.title}>
-                    <p>NUTFES BINGO {props.user}</p>
-                </div>
-                <div>
-                    <Modal buttonText="最新の番号を表示">
-                        <div className={styles.modalContent}>
-                            抽選された番号
-                            <p>25</p>
-                        </div>
-                    </Modal>
-                </div>
-            </div>
+            {children}
         </div>
-    )
-}
+    );
+};
 
 export default Header;
