+++ view-user/src/components/common/Modal/Modal.module.css
@@ -1,80 +1,83 @@
 .wrapper {
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 2;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 100%;
-    contain: content;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 2;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  contain: content;
 }
 
 .content {
-    position: relative;
-    z-index: 1;
-    box-sizing: border-box;
-    width: 25rem;
-    height: 25rem;
-    max-width: 700px;
-    max-height: 90vh;
-    padding: 40px;
-    overflow-y: auto;
-    background-color: #C67979;
-    box-shadow: 0 0 0 3.56rem #D9D9D9;
-    border-radius: 5px;
-    animation: zoomIn 0.8s cubic-bezier(0.25, 1, 0.5, 1) 1 forwards;
-    will-change: transform, opacity;
+  position: relative;
+  z-index: 1;
+  box-sizing: border-box;
+  object-fit: cover;
+  padding: 40px;
+  overflow: hidden;
+  background-color: #c67979;
+  box-shadow: 0 0 0 3.56rem #d9d9d9;
+  border-radius: 5px;
+  animation: zoomIn 0.8s cubic-bezier(0.25, 1, 0.5, 1) 1 forwards;
+  will-change: transform, opacity;
 }
 
-@keyframes zoomIn {
-    0% {
-        transform: scale(0.8);
-        opacity: 0;
-    }
-    100% {
-        opacity: 1;
-        transform: scale(1);
-    }
-}
-
-.btnOpen {
+.number {
     display: flex;
+    flex-flow: column;
     justify-content: center;
     align-items: center;
-    margin: 0 auto;
-    padding: .9em 2em;
-    border: 2px solid #fff;
-    border-radius: 5px;
-    box-shadow: 0 0 0 3px #333333;
-    background-color: #333333;
-    color: #fff;
-    font-weight: 600;
-    font-size: 1em;
+    color: #d9d9d9;
+    font-size: clamp(2rem, 2vw, 3rem);
+    font-weight: 700;
+    font-feature-settings: "palt" 1;
+  }
+
+  .number p {
+    display: flex;
+    flex-flow: column;
+    justify-content: center;
+    align-items: center;
+    color: #d9d9d9;
+    font-size: clamp(0.5rem, 50vw,16rem);
+    font-weight: 700;
+    font-feature-settings: "palt" 1;
+    line-height: 1;
+  }
+
+@keyframes zoomIn {
+  0% {
+    transform: scale(0.8);
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+    transform: scale(1);
+  }
 }
 
-
 .btnClose {
-    position: absolute;
-    top: 5px;
-    margin-top: -4px;
-    right: 10px;
-    padding: 0;
-    font-size: 2.7rem;
-    font-weight: bold;
-    line-height: 1;
-    color: #D9D9D9;
-    cursor: pointer;
-    background-color: transparent;
-    border: none;
+  position: absolute;
+  top: 3px;
+  right: 10px;
+  padding: 0;
+  font-size: 2.7rem;
+  font-weight: bold;
+  line-height: 1;
+  color: #d9d9d9;
+  cursor: pointer;
+  background-color: transparent;
+  border: none;
 }
 
 .background {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0 0 0 / 50%);
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0 0 0 / 50%);
 }
